Added tipbox and (unfinished) image uploading
The tipbox is currently at the toprightmost part of the interface.  This
is easily changed, however, it is there now because it would currently
overlap the user login link.  The image uploading is stubbed at the
moment as I am currently investigating tying everything together on the
backend, as well as adjusting asthetic issues with it, so it is not
completely finished