Fixed(?) tips div
I had to compromise at the moment.  After a long headache with the tips
div, I conceeded to simply place it beside the current body.  The issue
is twofold; first, there is an invisible border for the current
"content" div that I can't seem to remove.  This places a 5 pixel
separation between the two objects and eliminates my ability to fake the
tips box over an existing body.  Second, the content div has dynamic
heights.  This seems to only happen after login, but I don't want to
hardcode something like that for obvious reasons.  This can be changed
eventually, but I think its an ok compromise for now.  if not, it's
easily changed back.